requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2
aiosmtplib==3.0.1
jinja2==3.1.2
python-dotenv==1.0.0
slowapi==0.1.9

